+++ CustomItems/change-log.md
@@ -1,3 +1,9 @@
+# 0.0.6
+- Update to support new recipe format White Vinegar = `10x 166 "white wine"` which was suggested instead of `166 - "white wine" (10)`
+  - the mod "for now" will support both styles
+- fixed issue with variable recipes not using item sell prices to increase recipe sell price
+- Items and Recipes can now have id 0 and the mod will find the items (only modded items not base game items)
+
 # 0.0.5
 - added even more safety to the mod (bad headers in file)